Update test function
